refactor(CityDetail): create 2 new components that are now used in CityDetail
